# ApiGateway/Gateway/Gateway/Domain/User.cs
namespace Gateway.Domain
{
    public class User
    {
        public string cedula { get; set; }
        public string nombre { get; set; }
        public string apellido { get; set; }
        public string email { get; set; }
        public string contrasena { get; set; }
    }
}